requirements.txt
pyserial
PyBluez
Adafruit-Blinka
Adafruit-ADS1x15
